Enable to poll in either read or write mode
I have an EEPROM that want's to be polled in WRITE mode, i.e. R/W set to 0.
Currently, pyftdi only allows polling in READ mode. I added an extra parameter to the poll methods, so the user can choose to poll in WRITE mode.
I set the default of this parameter to READ mode, so the change is backwards-compatible.