Tristate SDA pin during ACK in I2C mode for non-drive-zero devices
This patch configures the SDA pin as input when the I2C slave puts
its N/ACK on the wire. This is so non "drive-zero" hardware (like
the 2232H) can be used to communicate with I2C devices without
hardware overhead. A current-limiting resistor on the I2C line is
still recommended (10-250 ohms, depending on your wire length and
supply voltage). This method is also used by FTDI themselves, for
example in their application note AN 411: "For the FT2232H and
FT4232H, the application includes additional GPIO writes to ADBUS
which tristate the AD1 Data Out pin when the line is not being
driven, to simulate an open-drain pin."