SPI losing bit
I'm using the CI MCP3008 (10-Bit A/D Converters) but I was always losing the fist bit (I got only 9-Bits) and got all bits inverted (0' becomes 1' and 1' becomes 0'). Although I have already used this CI with Arduino and Raspberry, it's fail on FT232H. 

After hour debugging,  I realize that the command to write  precede the command to read, thus the time between both command to be received and processed is very short but enough to lose one bit. By this way, the only way to solve this was send the command to FT device to start read and write at the same time. 

After solving the losing bit, the inverted bits problem disappear.

Probability it can help with mode 2 issue if we include one way to manage the initial polarity of the clock. (Unfortunately I don't have any device that operate with mode 2 to test )

Finally, the code should be revised to distinguish between read/write operation and write only operation.

Regards