Allow I2C controllers to be created for existing FTDI devices
If an FTDI device is already available, it is unnecessary to create it
again only for the purpose of instantiating an I2C controller. Allow an
FTDI device to be passed to the I2C controller's constructor and reuse
that FTDI device instead of opening a new one.

While at it, make the url argument of the I2cController.configure()
function optional for cases where the FTDI device was specified at the
time of construction.