Allow opening FTDI connections to specific USB devices
This provides a slightly lower-level mechanism for creating FTDI
connections by allowing a USB device to be directly passed. This
is useful if the application already gets these USB devices from
elsewhere and avoids having to fake up a URL (which could be not
very straightforward) just to match the already known device.