i2c: relax the bus after epilog in fake-tristate mode
set lines to high-z after epilog when fake-tristate is enabled

Similiar to issue #46, FT2232H/FT4232H never *really* relax the bus
after the epilog, because they don't support tristate in hardware. This
leads to blocking the bus when another master (besides the FTDI) is
present.

Implement fake-tristate for both SCL and SDA after the epilog to unblock
the bus.

Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>